Initial project strucure (#1)
* initial commit

* I have begun to write the graph module down on paper

* Fix clippy warnings so I can push.

I'm actually going to remove the pre-push hook, it's very obnoxious.
This should be done when trying to merge.

* Remove prepush hook that is just obnoxious.

Replace this with GitHub actions. People should be able to store and share their code on a remote
branch in whatever form they please.

* Begin on storage module with the promise of benchmarks and integration tests
Also removed cargo-husky from dev-dependencies.

Traits will be my saviour.